2025-02-25 OSS Sync (#15)
* Publish proposals and roadmap on the website (#1043)

* Allow Polaris CLI to set property values that contain '=' (#1003)

Currently the CLI parsing is too pessimistic and throws an error if
it detects '=' in the parsed value of a property. However, the split 1
semantics are already standard behavior for most parsers where the
right-hand-side is allowed to contain '='.

This commonly comes up if the value is itself a comma-separated kv list
or if it the value is a base64-encoded string.

* main: Update postgres Docker tag to v17.4 (#1049)

* main: Update dependency org.junit:junit-bom to v5.12.0 (#1045)

* Revert "Remove CallContext and its ThreadLocal usage" (#1000)

* revert b84f462

* resolve conflicts

* autolint

* autolint

* ...not possible to automatically add a synthetic no-args constructor to an unproxyable bean class.

* fix

* fixed non-quarkus build issues

* fixed non-quarkus build issues

* autolint

* more di issues

* add more producers

* add another producer

* autolint

* change callcontext usage in FileIOFactoryTest

* stablize tests

* autolint

* fix a test

* refactor TableCleanupTaskHandlerTest

* another fix, still npe

* autolint

* only task cleanup is problematic

* autolint

* Stabilize more tests

* autolint

* Avoid propagating old callcontexts across TestServices instances

* testservices update

* autolint

* Revert "autolint"

This reverts commit 66a7f22510582f3ebf71df7074d0d181ce26d2ac.

* Revert "testservices update"

This reverts commit 897a5a18cb2a366c53b68c3fba94292c35a2d0a5.

* changes to taskexecutorimpl

* autolint

---------

Co-authored-by: Dennis Huo <dennis.huo+oss@snowflake.com>

* Enable console color when running Polaris with Gradle (#977)

* CLI: add subcommand access for principals (#1019)

* Access subcommand access for principals

* Access subcommand access for principals

* Fix getting start spark (#1054)

* Use correct link format for adoc and update menu title for proposals (#1051)

* Add a debug log with the full exception (#1023)

* clean up

* spotless

* clean up

* refactor message

* no CVEM

* change log levels

* revert polaris-java dep change

* add testImplementation

* uppercase

* Remove all tightly coupled EntityCache dependencies in the main persistence stack (#1055)

Remove the EntityCacheEntry wrapper since the timestamp fields were never used anyways;
instead the underlying Caffeine cache transparently handles access times, and the
types of entries we cache are simply the existing ResolvedPolarisEntity.

Remove interactions of business logic with explicit "cache entries", instead
operating on ResolvedPolarisEntity. Fix the equals()/hashCode() behavior of
PolarisEntity to be compatible with PolarisBaseEntity as intended.

Improve code comments to explain the (current) relationship between PolarisEntity
and PolarisBaseEntity, and clarify the behaviors in Resolver.java.

Fully remove the PolarisRemoteCache interface and its methods. Add different
methods that aren't cache-specific instead.

* Remove PolarisMetaStoreSession from FileIOFactory/FileIOUtil in favor of CallContext (#1057)

This appeared to be some leaky divergence that occurred after CallContext had been
removed, but PolarisMetaStoreSession really is only a low-level implementation detail
that should never be handled by BasePolarisCatalog/FileIOFactory.

This plumbs CallContext explicitly into the FileIOFactory and FileIOUtil methods and
thus removes a large source of CallContext.getCurrentContext calls; now the threadlocal
doesn't have to be set at all in BasePolarisCatalogTest.

* Cleanup Iceberg Manifest list files on Purge (#1039)

* Fix small quickstart typo (#1061)

* Enable rat check on md files and fix the existing ones (#1050)

* main: Update dependency boto3 to v1.37.0 (#1052)

* main: Update dependency software.amazon.awssdk:bom to v2.30.27 (#1053)

* Integration tests: improve support for temporary folders (#987)

3 integration tests currently require a temporary folder for various purposes. The folders are generally hard-coded to some local filesystem location, which makes those tests impossible to execute against a remote Polaris instance.

This PR aims at fixing this situation by leveraging a new optional environment variable that points to a location that is accessible by both the server and the client running the tests. This would typically be an S3 bucket, for example, but could also be a shared volume mounted on both machines at the same mount point. If this environment variable is not present, then local execution is assumed and a temporary directory managed by JUnit is used instead.

* build-logic: exclude generated code from errorprone (#1035)

* build-logic: let javadoc depend on jandex (#1034)

* Add Prashant in the collaborators list (#1064)

* main: Update dependency org.slf4j:slf4j-api to v2.0.17 (#1065)

---------

Co-authored-by: JB Onofré <jbonofre@apache.org>
Co-authored-by: Dennis Huo <7410123+dennishuo@users.noreply.github.com>
Co-authored-by: Mend Renovate <bot@renovateapp.com>
Co-authored-by: Eric Maynard <eric.maynard+oss@snowflake.com>
Co-authored-by: Dennis Huo <dennis.huo+oss@snowflake.com>
Co-authored-by: Alexandre Dutra <adutra@users.noreply.github.com>
Co-authored-by: MonkeyCanCode <yongzheng0809@gmail.com>
Co-authored-by: Andrew Guterman <andrew.guterman1@gmail.com>
Co-authored-by: Prashant Singh <35593236+singhpk234@users.noreply.github.com>
Co-authored-by: Robert Stupp <snazy@snazy.de>